rename files and remove old import and export
